% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bined_converted_func.R
\name{bined_converted_func}
\alias{bined_converted_func}
\title{Title This is small function to output the final data as numerical type}
\usage{
bined_converted_func(converted_data, original_data, scale_numeric = FALSE)
}
\arguments{
\item{converted_data}{insert the converted data from the UFT_func}

\item{original_data}{insert the full original data}

\item{scale_numeric}{logical. Scale the data if True inserted. Default is False.}
}
\value{
a data frame the contain the tranformed data
}
\description{
Title This is small function to output the final data as numerical type
}
\examples{
X <- data.frame("ID" = as.numeric(c(1,2,3,4)), "Age" = c(21,15,44,14), "Sex" = c("Male", "Male", "Female", "Female" ),
"Name" = c("John","Samer", "Sara", "Carolina"), Smoking = c("Yes","No", "No","Yes"))
uft_x <- UFT_func(Data = X,Seed = 11)
bined_data <- bined_converted_func(converted_data = uft_x, original_data = X, scale_numeric = TRUE)
}
