% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_data.R
\name{Prepare_data}
\alias{Prepare_data}
\title{Title Prepare the data for the UFT_Func. Mainly, it converts factor and character type data to integer format.}
\usage{
Prepare_data(categorical_data)
}
\arguments{
\item{categorical_data}{is the original categorical data that the UFT_Func will transform to become numerical. It should be of data frame format.}
}
\value{
a data frame with all the data are of int type.
}
\description{
Title Prepare the data for the UFT_Func. Mainly, it converts factor and character type data to integer format.
}
\examples{
X <- data.frame("ID" = 1:4, "Sex" = c("Male", "Male", "Female", "Female" ), "Name" = c("John","Samer", "Sara", "Carolina"), Smoking = c("Yes","No", "No","Yes"))
Prepare_data(X)
}
