% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFT_func.R
\name{UFT_func}
\alias{UFT_func}
\title{Title Unsupervised Feature Transformation}
\usage{
UFT_func(Data, Seed)
}
\arguments{
\item{Data}{is the original data we want to deal with. It should be of data fram format. It may contain both categorical and numerical features.
The UFT_Func will split the data and drop the already numerical features.}

\item{Seed}{is for reproducibility reasons.}
}
\value{
The output from the UFT_Func is a dataframe of numerical columns replacing the original categorical features.
}
\description{
Title Unsupervised Feature Transformation
}
\examples{
X <- data.frame("ID" = as.numeric(c(1,2,3,4)), "Age" = c(21,15,44,14), "Sex" = c("Male", "Male", "Female", "Female" ), "Name" = c("John","Samer", "Sara", "Carolina"), Smoking = c("Yes","No", "No","Yes"))
}
