% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Splits.R
\name{Data_Classes_Split}
\alias{Data_Classes_Split}
\title{This is s small function the split the data according to thier types (numerica, integers or factors)}
\usage{
Data_Classes_Split(Data)
}
\arguments{
\item{Data}{is the dataframe we are spliting it's variables to different classes.}
}
\value{
return a list that each element contain the coloumns that belong to specific class
}
\description{
This is s small function the split the data according to thier types (numerica, integers or factors)
}
\details{
This function split the data to different classes depending on the data type that is contained in each variable in the data
}
\examples{
X <- data.frame("ID" = as.numeric(c(1,2,3,4)), "Age" = c(21,15,44,14), "Sex" = c("Male", "Male", "Female", "Female"),
"Name" = c("John","Samer", "Sara", "Carolina"), Smoking = c("Yes","No", "No","Yes"))
data_split <- Data_Classes_Split(X)

}
\author{
Rani Basna
}
